+++ src/cli.ts
@@ -1,5 +1,7 @@
 const VERSION = require('../package.json').version;
 
+import fs from 'fs';
+import path from 'path';
 import { program } from 'commander';
 
 import linguist from './index';
@@ -18,6 +20,7 @@ program
 	.option('-C|--childLanguages [bool]', 'Display child languages instead of their parents', false)
 	.option('-j|--json [bool]', 'Display the output as JSON', false)
 	.option('-t|--tree <traversal>', 'Which part of the output JSON to display (dot-delimited)')
+	.option('-F|--listFiles [bool]', 'Whether to list every matching file under the language results', false)
 	.option('-q|--quick [bool]', 'Skip complex language analysis (alias for -{A|I|H|S}=false)', false)
 	.option('-o|--offline [bool]', 'Use packaged data files instead of fetching latest from GitHub', false)
 	.option('-V|--keepVendored [bool]', 'Prevent skipping over vendored/generated files', false)
@@ -58,20 +61,50 @@ if (args.analyze) (async () => {
 		const sortedEntries = Object.entries(languages.results).sort((a, b) => a[1].bytes < b[1].bytes ? +1 : -1);
 		const totalBytes = languages.bytes;
 		console.log(`\n Analysed ${files.bytes.toLocaleString()} B from ${files.count} files with linguist-js`);
-		console.log(`\n Language analysis results:`);
+		console.log(`\n Language analysis results: \n`);
 		let count = 0;
 		if (sortedEntries.length === 0) console.log(`  None`);
+		// Collate files per language
+		const filesPerLanguage: Record<string, string[]> = {};
+		if (args.listFiles) {
+			for (const language of Object.keys(languages.results)) {
+				filesPerLanguage[language] = [];
+			}
+			for (const [file, lang] of Object.entries(files.results)) {
+				if (lang)
+					filesPerLanguage[lang].push(file);
+			}
+		}
 		// List parsed results
 		for (const [lang, { bytes, color }] of sortedEntries) {
+			const percent = (bytes: number) => bytes / (totalBytes || 1) * 100;
 			const fmtd = {
 				index: (++count).toString().padStart(2, ' '),
 				lang: lang.padEnd(24, ' '),
-				percent: (bytes / (totalBytes || 1) * 100).toFixed(2).padStart(5, ' '),
+				percent: percent(bytes).toFixed(2).padStart(5, ' '),
 				bytes: bytes.toLocaleString().padStart(10, ' '),
 				icon: colouredMsg(hexToRgb(color ?? '#ededed'), '\u2588'),
 			};
 			console.log(`  ${fmtd.index}. ${fmtd.icon} ${fmtd.lang} ${fmtd.percent}% ${fmtd.bytes} B`);
+
+			// If using `listFiles` option, list all files tagged as this language
+			if (args.listFiles) {
+				console.log(); // padding
+				for (const file of filesPerLanguage[lang]) {
+					let relFile = path.relative(path.resolve('.'), file).replace(/\\/g, '/');
+					if (!relFile.startsWith('../')) relFile = './' + relFile;
+					const bytes = (await fs.promises.stat(file)).size;
+					const fmtd2 = {
+						file: relFile.padEnd(42, ' '),
+						percent: percent(bytes).toFixed(2).padStart(5, ' '),
+						bytes: bytes.toLocaleString().padStart(10, ' '),
+					}
+					console.log(`        ${fmtd.icon} ${fmtd2.file}  ${fmtd2.percent}% ${fmtd2.bytes} B`);
+				}
+				console.log(); // padding
+			}
 		}
+		if (!args.listFiles) console.log(); // padding
 		console.log(` Total: ${totalBytes.toLocaleString()} B`);
 		// List unknown files/extensions
 		if (unknown.bytes > 0) {